# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/splash_screens": {
            "get": {
                "tags": [
                    "SplashScreens"
                ],
                "summary": "Retrieve a list of splash screens",
                "description": "Get a paginated list of splash screens",
                "operationId": "c651a0a6f9a6ee829d338d030211e3ca",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items to return per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved splash screens",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Splash screens retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SplashScreenSchema"
                                            }
                                        },
                                        "errors": {
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid parameters"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "SplashScreens"
                ],
                "summary": "Create a new splash screen",
                "description": "Creates a new splash screen with an image upload",
                "operationId": "7e79d5f8a7a1153b92f6ff1a54d42173",
                "requestBody": {
                    "description": "Create a new splash screen",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreSplashScreenRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully created splash screen",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Splash screen created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SplashScreenSchema"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid parameters"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/splash_screens/{id}": {
            "get": {
                "tags": [
                    "SplashScreens"
                ],
                "summary": "Retrieve a specific splash screen",
                "description": "Fetch the details of a specific splash screen by its ID.",
                "operationId": "9151da2e4f190ac36d8169e020d87d87",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the splash screen to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved splash screen",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Splash screen retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SplashScreenSchema"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Splash screen not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Splash screen not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "SplashScreens"
                ],
                "summary": "Delete a splash screen",
                "description": "Deletes the splash screen identified by the given ID.",
                "operationId": "2d8c7f647ecc9bcb581c271875feea5b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the splash screen to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Splash screen deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/SplashScreenSchema"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Splash screen deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found - The splash screen with the given ID does not exist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Splash screen not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid data or request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/splash_screens/update/{id}": {
            "post": {
                "tags": [
                    "SplashScreens"
                ],
                "summary": "Update an existing splash screen",
                "description": "Update the splash screen with the provided data. Only the fields provided in the request will be updated.",
                "operationId": "37928d2bc16b43f4c2a0466df83af250",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the splash screen to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSplashScreenRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Splash screen updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/SplashScreenSchema"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Splash screen updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Validation errors or missing required data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid input"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found - The splash screen with the given ID does not exist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Splash screen not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "StoreSplashScreenRequest": {
                "properties": {
                    "image": {
                        "description": "Splash screen image",
                        "type": "string",
                        "format": "binary"
                    },
                    "display_time": {
                        "description": "Display time for the splash screen",
                        "type": "integer",
                        "nullable": true
                    },
                    "text_message": {
                        "description": "Text message for the splash screen",
                        "type": "string"
                    },
                    "active": {
                        "description": "Whether the splash screen is active",
                        "type": "integer"
                    },
                    "display_order": {
                        "description": "Order to display the splash screen",
                        "type": "integer",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateSplashScreenRequest": {
                "properties": {
                    "image": {
                        "description": "Splash screen image (jpeg, png, jpg, gif, svg), max size 2MB",
                        "type": "string",
                        "format": "binary",
                        "nullable": true
                    },
                    "display_time": {
                        "description": "The display time in seconds for the splash screen",
                        "type": "integer",
                        "nullable": true
                    },
                    "text_message": {
                        "description": "Text message to display on the splash screen",
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "active": {
                        "description": "Indicates if the splash screen is active or not",
                        "type": "boolean",
                        "nullable": true
                    },
                    "display_order": {
                        "description": "The order in which the splash screen is displayed relative to others",
                        "type": "integer",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "SplashScreenSchema": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "image_url": {
                        "type": "string",
                        "example": "http://localhost:8000/storage/splash_images/Jph0KuayfVBC4szEsoFuGfzoIqJUhPnrJ93qkVCE.png"
                    },
                    "display_time": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "text_message": {
                        "type": "string",
                        "example": "lknnh"
                    },
                    "active": {
                        "type": "integer",
                        "example": 1
                    },
                    "display_order": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-24T13:16:20.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-24T13:22:16.000000Z"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "SplashScreens",
            "description": "SplashScreens"
        }
    ]
}